requirements.txt
PyQt5
PyYAML
watchdog
